Module Vhd_format.Element

type 'a t = [

| `Copy of 'a * int64 * int64	(*
Copy (t, offset, len) copies len sectors from sector offset from the file t
*)

| `Sectors of Cstruct.t	(*
a new sector (e.g. for metadata)
*)

| `Empty of int64	(*
empty space in sectors
*)

]

A disk can be streamed as a sequence of elements

val to_string : 'a t -> string

val len : 'a t -> int64

len t is the length of t in sectors