Module Vhd_format.F

val sector_size : int

Sector size of the vhd virtual disk, in bytes

val sector_shift : int

1 lsl sector_shift = sector_size

val cstruct_equal : Cstruct.t -> Cstruct.t -> bool

exception Invalid_sector of int64 * int64

An attempt to access (read/write) an invalid sector

val max_disk_size : int64

Maximum size of a dynamic disk in bytes

module UTF16 : sig ... end

module Feature : sig ... end

module Disk_type : sig ... end

module Host_OS : sig ... end

module Geometry : sig ... end

module Footer : sig ... end

module Platform_code : sig ... end

module Parent_locator : sig ... end

module Header : sig ... end

module BAT : sig ... end

module Batmap_header : sig ... end

module Batmap : sig ... end

module Bitmap : sig ... end

module Bitmap_cache : sig ... end

module Sector : sig ... end

module Vhd : sig ... end

module Raw : sig ... end

type size = {

total : int64;
(*
size of the final disk, in sectors
*)

metadata : int64;
(*
number of metadata sectors
*)

empty : int64;
(*
number of sectors of empty space
*)

copy : int64;
(*
number of copied sectors
*)

}

The amount of data contained in a stream, broken down by type

val empty : size

module Stream (A : S.ASYNC) : sig ... end

module Fragment : sig ... end

module From_input (I : S.INPUT) : sig ... end

module From_file (F : S.FILE) : sig ... end