Module F.BAT

type t

Absolute sector offset of a data block, where a data block contains a sector bitmap and then data

val equal : t -> t -> bool

equal t1 t2 is true if t1 and t2 represent the same data

val get : t -> int -> int32

get t i returns the ith entry

val set : t -> int -> int32 -> unit

set t i j sets the ith entry to j

val fold : (int -> int64 -> 'a -> 'a) -> t -> 'a -> 'a

fold f t initial folds f across all valid entries

val length : t -> int

length t the number of entries in the table

val to_string : t -> string

to_string t creates a debug-printable string