Module F.Bitmap

type t

val get : t -> int64 -> bool

get t sector is true if sector is present in the block

val set : t -> int64 -> (int64 * Cstruct.t) option

set t sector asserts the bit for sector, returning a (relative offset, data to be written to disk) pair.