Module F.From_file

Parameters

module F : S.FILE

Signature

module Vhd_IO : sig ... end

module Raw_IO : sig ... end

type 'a ll =

| Cons of 'a * unit -> 'a ll F.t

| End	(*
a lazy list
*)

val iter : ( 'a -> unit F.t ) -> 'a ll -> unit F.t

iter f stream applies each element from stream to f in order.

val fold_left : ( 'a -> 'b -> 'a F.t ) -> 'a -> 'b ll -> 'a F.t

fold_left f initial stream folds f across all the elements in the stream with neutral element initial

type 'a stream = {

elements : 'a Element.t ll;

size : size;

}

an image of a disk represented as a stream

val expand_empty : 'a stream -> 'a stream F.t

replaces 'Empty' elements with explicit writes of zeroes

val expand_copy : F.fd stream -> F.fd stream F.t

replaces 'Copy' elements with explicit writes of data

module Vhd_input : sig ... end

module Hybrid_input : sig ... end

module Raw_input : sig ... end