Module From_file.Vhd_input

val raw : ?from:F.fd Vhd.t -> F.fd Vhd.t -> F.fd stream F.t

raw t creates a raw-formatted stream representing the consolidated data present in the virtual disk t. If from is provided then the stream will contain the vhd differencing disk needed to transform from into t.

val vhd : ?from:F.fd Vhd.t -> ?emit_batmap:bool -> F.fd Vhd.t -> F.fd stream F.t

vhd t creates a vhd-formatted stream representing the consolidated data present in the virtual disk t. If from is provided then the stream will contain the vhd differencing disk needed to transform from into t. If ?emit_batmap is set then the resulting vhd will have the non-standard 'BATmap' metadata included.