Module F.Stream

Parameters

module A : S.ASYNC

Signature

type 'a ll =

| Cons of 'a * unit -> 'a ll A.t

| End	(*
a lazy list
*)

val iter : ( 'a -> unit A.t ) -> 'a ll -> unit A.t

iter f stream applies each element from stream to f in order.

val fold_left : ( 'a -> 'b -> 'a A.t ) -> 'a -> 'b ll -> 'a A.t

fold_left f initial stream folds f across all the elements in the stream with neutral element initial

type 'a stream = {

elements : 'a Element.t ll;

size : size;

}

an image of a disk represented as a stream