Module From_file.Hybrid_input

val raw : ?⁠from:F.fd Vhd.t -> F.fd -> F.fd Vhd.t -> F.fd stream F.t

raw ?from raw vhd creates a raw-formatted stream representing the consolidated data present in the virtual disk t, in terms of copies from the virtual disk raw. If from is provided then the stream will contain only the virtual updates required to transform from into t

val vhd : ?⁠from:F.fd Vhd.t -> F.fd -> F.fd Vhd.t -> F.fd stream F.t

vhd ?from raw vhd creates a vhd-formatted stream representing the consolidated data present in the virtual disk t, in terms of copies from the virtual disk raw. If from is provided then the stream will contain only the virtual updates required to transform from into t