Module F.Footer

type t = {

features : Feature.t list;

data_offset : int64;
(*
For dynamic and differencing disks, this is the absolute byte offset from the beginning of the file to the next structure. For fixed disks the value is undefined.
*)

time_stamp : int32;
(*
Creation time in seconds since midnight on January 1, 2000
*)

creator_application : string;
(*
Name of the application which created the image
*)

creator_version : int32;
(*
Version number of the application which created the image
*)

creator_host_os : Host_OS.t;

original_size : int64;
(*
size of the virtual disk in bytes at creation time
*)

current_size : int64;
(*
size of the virtual disk in bytes now
*)

geometry : Geometry.t;

disk_type : Disk_type.t;

checksum : int32;
(*
one's complement checksum of the footer with the checksum set to 0l
*)

uid : Uuidm.t;
(*
128-bit UUID
*)

saved_state : bool;
(*
true if the virtual machine is in a saved (suspended) state
*)

}

val create : ?features:Feature.t list -> data_offset:int64 -> ?time_stamp:int32 -> ?creator_application:string -> ?creator_version:int32 -> ?creator_host_os:Host_OS.t -> current_size:int64 -> ?original_size:int64 -> disk_type:Disk_type.t -> ?uid:Uuidm.t -> ?saved_state:bool -> unit -> t

val compute_checksum : t -> int32

compute the expected checksum value

val default_creator_application : string

val default_creator_version : int32

val sizeof : int

val marshal : Cstruct.t -> t -> t

val unmarshal : Cstruct.t -> (t, exn) Stdlib.result

val to_string : t -> string